refactor(cleanup): train_util.py の小ヘルパーと前処理クラスを再配置
- EPSILON / exists / default は参照ゼロのため完全削除（shim も置かない）
- ImageLoadingDataset を finetune/image_loading_dataset.py 新設に移動
  - 参照元 (finetune/prepare_buckets_latents.py / make_captions_by_git.py)
    の 2 箇所を新パス参照に書き換え
  - train_util shim は置かない（参照元が finetune/ 内に閉じているため）
- collator_class を library/dataset.py に移動し train_util から re-export
  - 学習スクリプト 15 ファイルの train_util.collator_class 参照は
    shim 経由でそのまま動作するため書き換え不要

train_util.py 295 → 222 行、pytest 150 passed
