+++ finetune/prepare_buckets_latents.py
@@ -20,6 +20,7 @@ import library.caching as caching
 import library.dataset as dataset
 import library.model_util as model_util
 import library.train_util as train_util
+from finetune.image_loading_dataset import ImageLoadingDataset
 from library.utils import setup_logging
 
 setup_logging()
@@ -119,7 +120,7 @@ def main(args):
 
     # 読み込みの高速化のためにDataLoaderを使うオプション
     if args.max_data_loader_n_workers is not None:
-        dataset = train_util.ImageLoadingDataset(image_paths)
+        dataset = ImageLoadingDataset(image_paths)
         data = torch.utils.data.DataLoader(
             dataset,
             batch_size=1,
